Finished

+++ exercises/23_conversions/using_as.rs
@@ -1,3 +1,11 @@
+/*
+ * @Author: ppz 2398672726@qq.com
+ * @Date: 2024-12-30 01:35:37
+ * @LastEditors: ppz 2398672726@qq.com
+ * @LastEditTime: 2025-01-07 22:04:28
+ * @FilePath: \rustlings\exercises\23_conversions\using_as.rs
+ * @Description: 这是默认设置,请设置`customMade`, 打开koroFileHeader查看配置 进行设置: https://github.com/OBKoro1/koro1FileHeader/wiki/%E9%85%8D%E7%BD%AE
+ */
 // Type casting in Rust is done via the usage of the `as` operator.
 // Note that the `as` operator is not only used when type casting. It also helps
 // with renaming imports.
@@ -5,7 +13,7 @@
 fn average(values: &[f64]) -> f64 {
     let total = values.iter().sum::<f64>();
     // TODO: Make a conversion before dividing.
-    total / values.len()
+    total / (values.len() as f64)
 }
 
 fn main() {
